GenomicAlignments 1.45.5: Deprecate CIGAR utilities and sequenceLayer()
The low-level CIGAR utilities and sequenceLayer() function are now in
their own package, the cigarillo package, but with different names.
Calling a low-level CIGAR utility function (or sequenceLayer()) defined
in GenomicAlignments still works but issues a deprecation warning that
indicates the name of the new function to use from cigarillo.

For example:

  > cigarOpTable(c("18M4S", "8M2X5M2I10M"))
       M I D N S H P = X
  [1,] 1 0 0 0 1 0 0 0 0
  [2,] 3 1 0 0 0 0 0 0 1
  Warning message:
  In call_new_fun_in_cigarillo("cigarOpTable", "tabulate_cigar_ops",  :
    cigarOpTable() is formally deprecated in GenomicAlignments >= 1.45.5
    and replaced with the tabulate_cigar_ops() function from the new
    cigarillo package